+++ R/verb-expand.R
@@ -33,16 +33,6 @@ expand.tbl_lazy <- function(data, ..., .name_repair = "check_unique") {
     abort("Must supply variables in `...`")
   }
 
-  extract_dot_vars <- function(.x) {
-    # ugly hack to deal with `nesting()`
-    if (quo_is_call(.x, name = "nesting")) {
-      x_expr <- quo_get_expr(.x)
-      call_args(x_expr)
-    } else {
-      list(quo_get_expr(.x))
-    }
-  }
-
   distinct_tbl_vars <- purrr::map(dots, extract_dot_vars)
 
   # now that `nesting()` has been unpacked resolve name conflicts
@@ -62,6 +52,23 @@ expand.tbl_lazy <- function(data, ..., .name_repair = "check_unique") {
   )
 
   purrr::reduce(distinct_tables, left_join, by = group_vars(data))
+}
+
+extract_dot_vars <- function(dot) {
+  # ugly hack to deal with `nesting()`
+  if (quo_is_call(dot, name = "nesting")) {
+    x_expr <- quo_get_expr(dot)
+    args <- call_args(x_expr)
+
+    repair <- args[[".name_repair"]] %||% "check_unique"
+    args[[".name_repair"]] <- NULL
+
+    args_named <- exprs_auto_name(args)
+    nms <- vctrs::vec_as_names(names(args_named), repair = repair)
+    exprs(!!!set_names(args_named, nms))
+  } else {
+    list(quo_get_expr(dot))
+  }
 }
 
 #' Complete a SQL table with missing combinations of data